Remote devices now send the raw ADC value for the battery level.

+++ timer_firmware/Battery.h
@@ -4,7 +4,7 @@
 void batteryInit();
 void batteryInit();
 void batterySampleTrigger();
-uint8_t batteryGetLevel();
+uint16_t batteryGetLevel();
 
 
 #endif /* BATTERY_H_ */
